op-service: rearrange blobs based on provided hashes
The behavior in the post-Fulu /blobs/ happy path and the
/blob_sidecars/ fallback slightly differed: in the former case,
returned blobs were ordered based their index in the blob, while in the
latter case they were ordered based on the provided hashes.

This bug was not caught in testing because all calls to BeaconBlobs
provide a slice of hashes sorted by index.